requirement.txt
streamlit
sentence-transformers
chromadb
langchain
langchain-community
openai
requests
scikit-learn
python-dotenv
